+++ src/hrdmc/estimators/__init__.py
@@ -1,11 +1,14 @@
 from hrdmc.estimators.density import estimate_density_profile
+from hrdmc.estimators.local_energy import LocalEnergyResult, estimate_local_energy
 from hrdmc.estimators.pair_distribution import PairDistributionResult, estimate_pair_distribution
 from hrdmc.estimators.structure_factor import StructureFactorResult, estimate_static_structure_factor
 
 __all__ = [
+    "LocalEnergyResult",
     "PairDistributionResult",
     "StructureFactorResult",
     "estimate_density_profile",
+    "estimate_local_energy",
     "estimate_pair_distribution",
     "estimate_static_structure_factor",
 ]